+++ src/tcerror.h
@@ -85,7 +85,7 @@ DEFINE_TCError(PipeLimit, "pipeline length reaches limit");
 DEFINE_TCError(ExpandLimit, "cannot expand too large path fragments");
 DEFINE_TCError(NullInPath, "found null characters in source path");
 DEFINE_TCError(NoGlobMatch, "no matches for glob pattern: `%s'");
-DEFINE_TCError(GlobRetLimit, "number of glob results reaches limit: `%s'");
+DEFINE_TCError(ExpandRetLimit, "number of expansion results reaches limit");
 DEFINE_TCError(NoGlobDir, "glob pattern always matches directory: `%s'");
 DEFINE_TCError(Constant, "must be constant expression");
 DEFINE_TCError(DupPattern, "duplicated pattern");